requirements.txt
# python>=3.8
numpy>1.21
h5py>3.7
hdf5plugin
pillow
matplotlib
nicegui>1.3.16